ECL-593: Addressed PR comments
